Initial setup for multiview development * MonthCalendar as subclass of Calendar class * methods updated to provide calendar and event data different periods (not only Month)
